Remove `filters` export from `@keystone-6/core/types`
