+++ .changeset/silent-dolphins-scream.md
@@ -1,0 +1,5 @@
+---
+'@keystone-6/core': minor
+---
+
+Removes `filters` export from `@keystone-6/core/types`
