+++ packages/core/src/fields/filters/providers/mysql.ts
@@ -1,9 +1,7 @@
 // Do not manually modify this file, it is automatically generated by the package at /prisma-utils in this repo.
 // Update the script if you need this file to be different
 
-import { graphql } from '../../schema';
-
-import { QueryMode } from '../../next-fields';
+import { graphql } from '../../../types/schema';
 
 type StringNullableFilterType = graphql.InputObjectType<{
   // can be null
@@ -19,7 +17,6 @@ type StringNullableFilterType = graphql.InputObjectType<{
   contains: graphql.Arg<typeof graphql.String>;
   startsWith: graphql.Arg<typeof graphql.String>;
   endsWith: graphql.Arg<typeof graphql.String>;
-  mode: graphql.Arg<typeof QueryMode>;
   // can be null
   not: graphql.Arg<typeof NestedStringNullableFilter>;
 }>;
@@ -40,7 +37,6 @@ const StringNullableFilter: StringNullableFilterType = graphql.inputObject({
     contains: graphql.arg({ type: graphql.String }),
     startsWith: graphql.arg({ type: graphql.String }),
     endsWith: graphql.arg({ type: graphql.String }),
-    mode: graphql.arg({ type: QueryMode }),
     // can be null
     not: graphql.arg({ type: NestedStringNullableFilter }),
   }),
@@ -96,7 +92,6 @@ type StringFilterType = graphql.InputObjectType<{
   contains: graphql.Arg<typeof graphql.String>;
   startsWith: graphql.Arg<typeof graphql.String>;
   endsWith: graphql.Arg<typeof graphql.String>;
-  mode: graphql.Arg<typeof QueryMode>;
   not: graphql.Arg<typeof NestedStringFilter>;
 }>;
 
@@ -113,7 +108,6 @@ const StringFilter: StringFilterType = graphql.inputObject({
     contains: graphql.arg({ type: graphql.String }),
     startsWith: graphql.arg({ type: graphql.String }),
     endsWith: graphql.arg({ type: graphql.String }),
-    mode: graphql.arg({ type: QueryMode }),
     not: graphql.arg({ type: NestedStringFilter }),
   }),
 });
@@ -149,29 +143,6 @@ const NestedStringFilter: NestedStringFilterType = graphql.inputObject({
   }),
 });
 
-type StringNullableListFilterType = graphql.InputObjectType<{
-  // can be null
-  equals: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.String>>>;
-  // can be null
-  has: graphql.Arg<typeof graphql.String>;
-  hasEvery: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.String>>>;
-  hasSome: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.String>>>;
-  isEmpty: graphql.Arg<typeof graphql.Boolean>;
-}>;
-
-const StringNullableListFilter: StringNullableListFilterType = graphql.inputObject({
-  name: 'StringNullableListFilter',
-  fields: () => ({
-    // can be null
-    equals: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.String)) }),
-    // can be null
-    has: graphql.arg({ type: graphql.String }),
-    hasEvery: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.String)) }),
-    hasSome: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.String)) }),
-    isEmpty: graphql.arg({ type: graphql.Boolean }),
-  }),
-});
-
 type BoolNullableFilterType = graphql.InputObjectType<{
   // can be null
   equals: graphql.Arg<typeof graphql.Boolean>;
@@ -199,29 +170,6 @@ const BoolFilter: BoolFilterType = graphql.inputObject({
   fields: () => ({
     equals: graphql.arg({ type: graphql.Boolean }),
     not: graphql.arg({ type: BoolFilter }),
-  }),
-});
-
-type BoolNullableListFilterType = graphql.InputObjectType<{
-  // can be null
-  equals: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Boolean>>>;
-  // can be null
-  has: graphql.Arg<typeof graphql.Boolean>;
-  hasEvery: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Boolean>>>;
-  hasSome: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Boolean>>>;
-  isEmpty: graphql.Arg<typeof graphql.Boolean>;
-}>;
-
-const BoolNullableListFilter: BoolNullableListFilterType = graphql.inputObject({
-  name: 'BooleanNullableListFilter',
-  fields: () => ({
-    // can be null
-    equals: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Boolean)) }),
-    // can be null
-    has: graphql.arg({ type: graphql.Boolean }),
-    hasEvery: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Boolean)) }),
-    hasSome: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Boolean)) }),
-    isEmpty: graphql.arg({ type: graphql.Boolean }),
   }),
 });
 
@@ -283,29 +231,6 @@ const IntFilter: IntFilterType = graphql.inputObject({
   }),
 });
 
-type IntNullableListFilterType = graphql.InputObjectType<{
-  // can be null
-  equals: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Int>>>;
-  // can be null
-  has: graphql.Arg<typeof graphql.Int>;
-  hasEvery: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Int>>>;
-  hasSome: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Int>>>;
-  isEmpty: graphql.Arg<typeof graphql.Boolean>;
-}>;
-
-const IntNullableListFilter: IntNullableListFilterType = graphql.inputObject({
-  name: 'IntNullableListFilter',
-  fields: () => ({
-    // can be null
-    equals: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Int)) }),
-    // can be null
-    has: graphql.arg({ type: graphql.Int }),
-    hasEvery: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Int)) }),
-    hasSome: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Int)) }),
-    isEmpty: graphql.arg({ type: graphql.Boolean }),
-  }),
-});
-
 type FloatNullableFilterType = graphql.InputObjectType<{
   // can be null
   equals: graphql.Arg<typeof graphql.Float>;
@@ -361,29 +286,6 @@ const FloatFilter: FloatFilterType = graphql.inputObject({
     gt: graphql.arg({ type: graphql.Float }),
     gte: graphql.arg({ type: graphql.Float }),
     not: graphql.arg({ type: FloatFilter }),
-  }),
-});
-
-type FloatNullableListFilterType = graphql.InputObjectType<{
-  // can be null
-  equals: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Float>>>;
-  // can be null
-  has: graphql.Arg<typeof graphql.Float>;
-  hasEvery: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Float>>>;
-  hasSome: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Float>>>;
-  isEmpty: graphql.Arg<typeof graphql.Boolean>;
-}>;
-
-const FloatNullableListFilter: FloatNullableListFilterType = graphql.inputObject({
-  name: 'FloatNullableListFilter',
-  fields: () => ({
-    // can be null
-    equals: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Float)) }),
-    // can be null
-    has: graphql.arg({ type: graphql.Float }),
-    hasEvery: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Float)) }),
-    hasSome: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Float)) }),
-    isEmpty: graphql.arg({ type: graphql.Boolean }),
   }),
 });
 
@@ -445,29 +347,6 @@ const DateTimeFilter: DateTimeFilterType = graphql.inputObject({
   }),
 });
 
-type DateTimeNullableListFilterType = graphql.InputObjectType<{
-  // can be null
-  equals: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.DateTime>>>;
-  // can be null
-  has: graphql.Arg<typeof graphql.DateTime>;
-  hasEvery: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.DateTime>>>;
-  hasSome: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.DateTime>>>;
-  isEmpty: graphql.Arg<typeof graphql.Boolean>;
-}>;
-
-const DateTimeNullableListFilter: DateTimeNullableListFilterType = graphql.inputObject({
-  name: 'DateTimeNullableListFilter',
-  fields: () => ({
-    // can be null
-    equals: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.DateTime)) }),
-    // can be null
-    has: graphql.arg({ type: graphql.DateTime }),
-    hasEvery: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.DateTime)) }),
-    hasSome: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.DateTime)) }),
-    isEmpty: graphql.arg({ type: graphql.Boolean }),
-  }),
-});
-
 type DecimalNullableFilterType = graphql.InputObjectType<{
   // can be null
   equals: graphql.Arg<typeof graphql.Decimal>;
@@ -523,29 +402,6 @@ const DecimalFilter: DecimalFilterType = graphql.inputObject({
     gt: graphql.arg({ type: graphql.Decimal }),
     gte: graphql.arg({ type: graphql.Decimal }),
     not: graphql.arg({ type: DecimalFilter }),
-  }),
-});
-
-type DecimalNullableListFilterType = graphql.InputObjectType<{
-  // can be null
-  equals: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Decimal>>>;
-  // can be null
-  has: graphql.Arg<typeof graphql.Decimal>;
-  hasEvery: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Decimal>>>;
-  hasSome: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.Decimal>>>;
-  isEmpty: graphql.Arg<typeof graphql.Boolean>;
-}>;
-
-const DecimalNullableListFilter: DecimalNullableListFilterType = graphql.inputObject({
-  name: 'DecimalNullableListFilter',
-  fields: () => ({
-    // can be null
-    equals: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Decimal)) }),
-    // can be null
-    has: graphql.arg({ type: graphql.Decimal }),
-    hasEvery: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Decimal)) }),
-    hasSome: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.Decimal)) }),
-    isEmpty: graphql.arg({ type: graphql.Boolean }),
   }),
 });
 
@@ -607,69 +463,39 @@ const BigIntFilter: BigIntFilterType = graphql.inputObject({
   }),
 });
 
-type BigIntNullableListFilterType = graphql.InputObjectType<{
-  // can be null
-  equals: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.BigInt>>>;
-  // can be null
-  has: graphql.Arg<typeof graphql.BigInt>;
-  hasEvery: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.BigInt>>>;
-  hasSome: graphql.Arg<graphql.ListType<graphql.NonNullType<typeof graphql.BigInt>>>;
-  isEmpty: graphql.Arg<typeof graphql.Boolean>;
-}>;
-
-const BigIntNullableListFilter: BigIntNullableListFilterType = graphql.inputObject({
-  name: 'BigIntNullableListFilter',
-  fields: () => ({
-    // can be null
-    equals: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.BigInt)) }),
-    // can be null
-    has: graphql.arg({ type: graphql.BigInt }),
-    hasEvery: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.BigInt)) }),
-    hasSome: graphql.arg({ type: graphql.list(graphql.nonNull(graphql.BigInt)) }),
-    isEmpty: graphql.arg({ type: graphql.Boolean }),
-  }),
-});
-
 export const String = {
   optional: StringNullableFilter,
   required: StringFilter,
-  many: StringNullableListFilter,
 };
 
 export const Boolean = {
   optional: BoolNullableFilter,
   required: BoolFilter,
-  many: BoolNullableListFilter,
 };
 
 export const Int = {
   optional: IntNullableFilter,
   required: IntFilter,
-  many: IntNullableListFilter,
 };
 
 export const Float = {
   optional: FloatNullableFilter,
   required: FloatFilter,
-  many: FloatNullableListFilter,
 };
 
 export const DateTime = {
   optional: DateTimeNullableFilter,
   required: DateTimeFilter,
-  many: DateTimeNullableListFilter,
 };
 
 export const Decimal = {
   optional: DecimalNullableFilter,
   required: DecimalFilter,
-  many: DecimalNullableListFilter,
 };
 
 export const BigInt = {
   optional: BigIntNullableFilter,
   required: BigIntFilter,
-  many: BigIntNullableListFilter,
 };
 
 export { enumFilters as enum } from '../enum-filter';
