+++ packages/core/src/fields/filters/providers/sqlite.ts
@@ -1,7 +1,7 @@
 // Do not manually modify this file, it is automatically generated by the package at /prisma-utils in this repo.
 // Update the script if you need this file to be different
 
-import { graphql } from '../../schema';
+import { graphql } from '../../../types/schema';
 
 type StringNullableFilterType = graphql.InputObjectType<{
   // can be null
