+++ packages/core/src/fields/types/float/index.ts
@@ -6,7 +6,6 @@ import {
   CommonFieldConfig,
   fieldType,
   orderDirectionEnum,
-  filters,
 } from '../../../types';
 import { graphql } from '../../..';
 import {
@@ -14,6 +13,7 @@ import {
   assertReadIsNonNullAllowed,
   getResolvedIsNullable,
 } from '../../non-null-graphql';
+import { filters } from '../../filters';
 
 export type FloatFieldConfig<ListTypeInfo extends BaseListTypeInfo> =
   CommonFieldConfig<ListTypeInfo> & {
