+++ packages/core/src/fields/types/timestamp/index.ts
@@ -5,7 +5,6 @@ import {
   FieldTypeFunc,
   CommonFieldConfig,
   orderDirectionEnum,
-  filters,
 } from '../../../types';
 import { graphql } from '../../..';
 import {
@@ -13,6 +12,7 @@ import {
   assertReadIsNonNullAllowed,
   getResolvedIsNullable,
 } from '../../non-null-graphql';
+import { filters } from '../../filters';
 import { TimestampFieldMeta } from './views';
 
 export type TimestampFieldConfig<ListTypeInfo extends BaseListTypeInfo> =
