+++ prisma-utils/src/index.ts
@@ -97,20 +97,12 @@ ${
       );
     }
 
-    await fs.outputFile(
-      `${__dirname}/generated/only-filters/${provider}.json`,
-      JSON.stringify(
-        Object.fromEntries([...referencedTypes].map(typeName => [typeName, types[typeName]])),
-        null,
-        2
-      )
-    );
-    const filepath = `${__dirname}/../../packages/core/src/types/filters/providers/${provider}.ts`;
+    const filepath = `${__dirname}/../../packages/core/src/fields/filters/providers/${provider}.ts`;
     const newContent = format(
       `// Do not manually modify this file, it is automatically generated by the package at /prisma-utils in this repo.
 // Update the script if you need this file to be different
       
-      import { graphql } from '../../schema';
+import { graphql } from '../../../types/schema'
 
 
 ${provider === 'postgresql' ? `import { QueryMode } from '../../next-fields'` : ''}
